requirements.txt
torch==2.8.0
torchvision==0.23.0
matplotlib==3.10.5
scikit-learn==1.7.1
numpy==2.2.1
pillow==11.2.1